% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cepiigravity-package.R
\docType{package}
\name{cepiigravity-package}
\alias{cepiigravity}
\alias{cepiigravity-package}
\title{cepiigravity: The CEPII Gravity Database}
\description{
The Gravity database aims at gathering in a single place a set of variables that could be useful to researchers or practitioners willing to understand the determinants of international trade. Each observation corresponds to a combination of exporter-importer-year (i.e. origin-destination-year), for which we provide trade flows, as well as geographic, cultural, trade facilitation and macroeconomic variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/cepiigravity}
  \item Report bugs at \url{https://github.com/pachamaltese/cepiigravity/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{hello+r@pacha.dev} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Other contributors:
\itemize{
  \item  Centre d'études prospectives et d'informations internationales (CEPII) [data contributor]
}

}
\keyword{internal}
