% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{cepiigravity_connect}
\alias{cepiigravity_connect}
\title{Connect to the gravity database}
\usage{
cepiigravity_connect(dir = cepiigravity_path())
}
\arguments{
\item{dir}{Database location on disk. Defaults to the \code{cepiigravity}
directory inside the R user folder or the \code{GRAVITY_DIR} environment variable
if specified.}
}
\description{
Returns a local database connection. This corresponds to a DBI-compatible
DuckDB database.
}
\examples{
\dontrun{
 DBI::dbListTables(gravity_connect())

 DBI::dbGetQuery(
  gravity_connect(),
  'SELECT * FROM gravity WHERE year = 2010'
 )
}
}
