% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{cepiigravity_delete}
\alias{cepiigravity_delete}
\title{Delete the gravity database from your computer}
\usage{
cepiigravity_delete(ask = TRUE)
}
\arguments{
\item{ask}{If so, a menu will be displayed to confirm the action to
delete any existing gravity database. By default it is true.}
}
\description{
Deletes the \code{cepiigravity} directory and all of its contents, including
all versions of the gravity database created with any DuckDB version.
}
\examples{
\dontrun{ gravity_delete() }
}
