% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{cepiigravity_disconnect}
\alias{cepiigravity_disconnect}
\title{Disconnect the gravity database}
\usage{
cepiigravity_disconnect()
}
\description{
An auxiliary function to disconnect from the database.
}
\examples{
cepiigravity_disconnect()
}
