% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{gravity_disconnect}
\alias{gravity_disconnect}
\title{Disconnect the gravity database}
\usage{
gravity_disconnect()
}
\description{
An auxiliary function to disconnect from the database.
}
\examples{
gravity_disconnect()
}
